Add files via upload
In this document, we enter a critical phase of the Environment monitoring ( Indoor Air Quality Monitoring project) , focusing on the development of the code that forms the system's backbone. Through meticulous examination, we will unveil the intricacies of the code's architecture and delve into its core functions. Furthermore, we will provide precise instructions for executing the code to drive the monitoring system effectively. This document serves as an essential reference, tailored for a discerning audience, including project stakeholders, developers, and those with a vested interest in comprehending the intricate technical foundations of our pioneering monitoring system.