Add files via upload
This document is all about our indoor air quality monitoring project. We've covered the basics, including an introduction to why we did it. You'll find the hardware components listed, the source code on Wokwi for you to check out, a simulation, and finally, our conclusion. In phase 4, the platform development and working of the project is included